+++ core/tests/integration_tests/mod.rs
@@ -21,6 +21,7 @@ mod test_cmd_cat;
 mod test_cmd_clean;
 mod test_cmd_completion;
 mod test_cmd_diff;
+mod test_cmd_dump;
 mod test_cmd_find;
 mod test_cmd_forget;
 mod test_cmd_init;